requirements.txt
fastapi
uvicorn
newspaper3k
textblob
nltk
